+++ lib/ts/axios.ts
@@ -23,7 +23,8 @@ import AuthHttpRequestFetch, {
     onInvalidClaimResponse,
     setToken,
     fireSessionUpdateEventsIfNecessary,
-    getTokenForHeaderAuth
+    getTokenForHeaderAuth,
+    updateClockSkewUsingFrontToken
 } from "./fetch";
 import { PROCESS_STATE, ProcessState } from "./processState";
 import WindowHandlerReference from "./utils/windowHandler";
@@ -556,6 +557,15 @@ async function saveTokensFromHeaders(response: AxiosResponse) {
     if (frontToken !== undefined) {
         logDebugMessage("doRequest: Setting sFrontToken: " + frontToken);
         await FrontToken.setItem(frontToken);
+
+        // Converting axios headers to fetch headers to pass to updateClockSkewUsingFrontToken
+        const responseHeaders = new Headers();
+        Object.entries(response.headers).forEach(([key, value]) => {
+            Array.isArray(value)
+                ? value.forEach(item => responseHeaders.append(key, item))
+                : responseHeaders.append(key, value);
+        });
+        updateClockSkewUsingFrontToken({ frontToken, responseHeaders });
     }
 
     const antiCsrfToken = response.headers["anti-csrf"];