+++ lib/ts/types.ts
@@ -17,6 +17,7 @@ import OverrideableBuilder from "supertokens-js-override";
 import { CookieHandlerInput } from "./utils/cookieHandler/types";
 import { WindowHandlerInput } from "./utils/windowHandler/types";
 import { LockFactory } from "./utils/lockFactory/types";
+import { DateProviderInput } from "./utils/dateProvider/types";
 
 export type Event =
     | {
@@ -57,6 +58,7 @@ export type InputType = {
     sessionTokenBackendDomain?: string;
     cookieHandler?: CookieHandlerInput;
     windowHandler?: WindowHandlerInput;
+    dateProvider?: DateProviderInput;
     preAPIHook?: RecipePreAPIHookFunction;
     postAPIHook?: RecipePostAPIHookFunction;
     onHandleEvent?: EventHandler;
@@ -150,6 +152,8 @@ export type RecipeInterface = {
         apiDomain: string,
         sessionTokenBackendDomain: string | undefined
     ): boolean;
+
+    getClockSkewInMillis(params: { accessTokenPayload: any; responseHeaders: Headers }): number;
 };
 
 export type ClaimValidationResult = { isValid: true } | { isValid: false; reason?: any };